requirements.txt
torch==1.9.1
torchvision==0.10.1
Pillow==8.2.0
pdf2image==1.16.0
fire==0.4.0
lmdb==1.2.1
pytz==2021.1